moved validation into metadata module
