+++ crates/fluvio-connector-package/src/lib.rs
@@ -1,4 +1,2 @@
 pub mod metadata;
 pub mod config;
-
-pub mod validation;
